pyproject.toml
[project]
name = "sports-sync-x"
version = "0.1.0"
description = "sync sport between garmin and coros"
requires-python = ">=3.14"
dependencies = [
    "black>=26.1.0",
    "boto3>=1.42.48",
    "boto3-stubs~=1.42.49",
    "garth>=0.6.3",
    "logging>=0.4.9.6",
    "oss2>=2.19.1",
    "pyrefly>=0.52.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
    "sqlmodel>=0.0.33",
    "ty>=0.0.17",
    "urllib3>=2.6.3",
]
